:star2: Add channel share link
