+++ twake/frontend/src/app/features/global/utils/CopyClipboard.ts
@@ -1,0 +1,8 @@
+export function copyToClipboard(url: string): void {
+    const el = document.createElement('textarea');
+    el.value = url;
+    document.body.appendChild(el);
+    el.select();
+    document.execCommand('copy');
+    document.body.removeChild(el);
+}
