+++ twake/frontend/src/app/views/client/channels-bar/Parts/Channel/ChannelMenu.tsx
@@ -18,7 +18,7 @@ import LockedGuestsPopup from 'app/components/locked-features-components/locked-
 import InitService from 'app/features/global/services/init-service';
 import ChannelsMineAPIClient from 'app/features/channels/api/channels-mine-api-client';
 import ChannelMembersAPIClient from 'app/features/channel-members/api/channel-members-api-client';
-import { isDirectChannel, isPrivateChannel } from 'app/features/channels/utils/utils';
+import { isDirectChannel, isPrivateChannel, isPublicChannel } from 'app/features/channels/utils/utils';
 import { useCurrentUser } from 'app/features/users/hooks/use-current-user';
 import useRouterWorkspace from 'app/features/router/hooks/use-router-workspace';
 import { ToasterService as Toaster } from 'app/features/global/services/toaster-service';
@@ -28,6 +28,7 @@ import ChannelAPIClient from 'app/features/channels/api/channel-api-client';
 import { useRefreshDirectChannels } from 'app/features/channels/hooks/use-direct-channels';
 import { useChannelNotifications } from 'app/features/users/hooks/use-notifications';
 import consoleService from 'app/features/console/services/console-service';
+import { copyToClipboard } from 'app/features/global/utils/CopyClipboard';
 
 type PropsType = {
   channel: ChannelType;
@@ -206,6 +207,20 @@ const FullMenu = (props: PropsType): JSX.Element => {
       },
     },
     {
+      type: 'menu',
+      text: Languages.t('scenes.app.channelsbar.channel_copy_link'),
+      hide: !(props.channel.visibility && isPublicChannel(props.channel.visibility)),
+      onClick: () => {
+        const url = `${document.location.origin}${RouterServices.generateRouteFromState({
+          workspaceId: props.channel.workspace_id || '',
+          companyId: props.channel.company_id,
+          channelId: props.channel.id
+        })}`;
+
+        copyToClipboard(url);
+      },
+    },
+    {
       hide: !(
         AccessRightsService.hasLevel(workspaceId, 'member') &&
         AccessRightsService.getCompanyLevel(companyId) !== 'guest'
@@ -282,7 +297,7 @@ const FullMenu = (props: PropsType): JSX.Element => {
       ],
     });
     menu.splice(
-      4,
+      5,
       0,
       {
         type: 'menu',
